[Chore] LIVD-415 - 폴더링
