+++ Projects/LivithNetworking/Sources/Foundation/Client/NetworkError.swift
