--- Projects/LivithNetworking/Sources/Request/HTTPMethod.swift
@@ -1,15 +1,0 @@
-//
-//  HTTPMethod.swift
-//  LivithNetworking
-//
-//  Created by 김진웅 on 5/8/26.
-//  Copyright © 2026 Livith. All rights reserved.
-//
-
-public enum HTTPMethod: String {
-    case get = "GET"
-    case post = "POST"
-    case put = "PUT"
-    case patch = "PATCH"
-    case delete = "DELETE"
-}
